Added multilighting with multipass and additive blending.

+++ bge/src/driver/gl3.h
@@ -92,7 +92,7 @@ class GL3 : public AbstractDriver
     QList<Light> m_lights;
     QHash<QString, Storage::Material*> m_materials;
     static const quint8 m_maxLights = 4;
-    quint8 m_renderedLights;
+    quint32 m_renderedLights;
 
     void load(Storage::Mesh* mesh);
     void load(Storage::ShaderProgram* shaderProgram);
